app/requirements.txt
Flask
flask-mysqldb
mysql-connector
mysql-client
